move layout out of app
