+++ web_frontend/src/main.tsx
@@ -2,9 +2,13 @@ import "regenerator-runtime";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import { RecoilRoot } from "recoil";
+import { MantineProvider } from "@mantine/core";
+import { theme } from "./theme.ts";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <RecoilRoot>
-    <App />
+    <MantineProvider withGlobalStyles withNormalizeCSS theme={theme}>
+      <App />
+    </MantineProvider>
   </RecoilRoot>
 );
